+++ ruoyi-app/src/main/java/com/ruoyi/app/service/impl/AppMessageServiceImpl.java
@@ -1,6 +1,8 @@
 package com.ruoyi.app.service.impl;
 
 import java.util.List;
+
+import com.ruoyi.common.utils.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.ruoyi.app.mapper.AppMessageMapper;
@@ -52,6 +54,7 @@ public class AppMessageServiceImpl implements IAppMessageService
     @Override
     public int insertAppMessage(AppMessage appMessage)
     {
+        appMessage.setUid(SecurityUtils.getUserId());
         return appMessageMapper.insertAppMessage(appMessage);
     }
 
@@ -88,6 +91,9 @@ public class AppMessageServiceImpl implements IAppMessageService
     @Override
     public int deleteAppMessageById(Long id)
     {
-        return appMessageMapper.deleteAppMessageById(id);
+        if (selectAppMessageById(id)!=null&&selectAppMessageById(id).getUid().equals(SecurityUtils.getUserId())){
+            return appMessageMapper.deleteAppMessageById(id);
+        }
+        return 0;
     }
 }
